requirements.txt
sentence-transformers>=2.2.0
scikit-learn>=1.0.2
flask>=2.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
spacy>=3.5.0
numpy>=1.22.0
pyyaml>=6.0
python-dotenv>=0.19.0
flask-cors>=3.0.10
email-validator>=1.1.3
phonenumbers>=8.12.0
pandas>=1.4.0
matplotlib>=3.5.0
seaborn>=0.11.0